Direct FS test cases were fixed

+++ backend/tests/utils/fs/direct.py
@@ -14,23 +14,25 @@ class DirectTestCase(unittest.TestCase):
         'second.md': 'Data of second file'
     }
 
+    files_to_cleanup = []
+
     def setUp(self):
         self.container_path = os.path.join(app.root_path, '../tmp')
         self.fs = Direct(self.container_path)
         for (path, data) in self.mock_files.iteritems():
             self.create_file(path, data)
+            self.files_to_cleanup.append(path)
 
     def tearDown(self):
-        listing = [f for f in os.listdir(self.container_path) if
-                   os.path.isfile(os.path.join(self.container_path, f))]
-        for path in listing:
-            self.remove_file(path)
+        for path in self.files_to_cleanup:
+            file_path = os.path.join(self.container_path, path)
+            self.remove_file(file_path)
 
     def test_list(self):
         result = self.fs.list()
 
         self.assertIsInstance(result, list)
-        self.assertEqual(len(result), 2)
+        self.assertEqual(len(result), 3)
 
     def test_read(self):
         result = self.fs.read('first.md')
@@ -47,14 +49,18 @@ class DirectTestCase(unittest.TestCase):
             return None
         except Exception:
             self.fail("We wait an NotExistsException exception")
-            return None
 
         self.fail("We wait an exception")
 
     def test_write_new(self):
-        result = self.fs.write('third.md', 'Data of third file')
+        path = 'third.md'
+        data = 'Data of third file'
 
-        self.assertEqual(result, 'Data of third file')
+        result = self.fs.write(path, data)
+
+        self.files_to_cleanup.append(path)
+
+        self.assertEqual(result, data)
 
     def test_write_existing(self):
         path = 'second.md'
@@ -75,7 +81,7 @@ class DirectTestCase(unittest.TestCase):
         listing = [f for f in os.listdir(self.container_path) if
                    os.path.isfile(os.path.join(self.container_path, f))]
 
-        self.assertEqual(len(listing), 1)
+        self.assertEqual(len(listing), 2)
 
     def test_remove_unknown(self):
         path = 'unknown.md'
@@ -86,15 +92,19 @@ class DirectTestCase(unittest.TestCase):
             return None
         except Exception:
             self.fail("We wait an NotExistsException exception")
-            return None
 
         self.fail("We wait an exception")
 
     def create_file(self, path, data):
         file_path = os.path.join(self.container_path, path)
+
         with open(file_path, 'w') as fh:
             fh.write(data)
 
     def remove_file(self, path):
         file_path = os.path.join(self.container_path, path)
+
+        if not os.path.isfile(os.path.join(self.container_path, file_path)):
+            return None
+
         os.remove(file_path)
